refining_Data/Classify/requirements.txt
absl-py==0.13.0
astunparse==1.6.3
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
et-xmlfile==1.1.0
flatbuffers==1.12
gast==0.4.0
google-auth==1.32.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
idna==2.10
joblib==1.0.1
Keras==2.4.3
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kss==2.5.1
Markdown==3.3.4
nltk==3.6.2
numpy==1.19.5
oauthlib==3.1.1
openpyxl==3.0.7
opt-einsum==3.3.0
pandas==1.3.0
protobuf==3.17.3
psutil==5.8.0
py-hanspell==1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pykospacing @ git+https://github.com/haven-jeon/PyKoSpacing.git@7bd81fe9c55a54e1b8e6f6bd5931b87bec98784f
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
regex==2021.7.6
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.7.0
six==1.15.0
soynlp==0.0.493
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
threadpoolctl==2.2.0
tqdm==4.61.2
typing-extensions==3.7.4.3
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
